requirements.txt
aiogram==3.4.0
asyncpg==0.29.0
sqlalchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9